pyproject.toml
[tool.poetry]
name = "book-worm"
version = "0.1.0"
description = "Telegram-bot for managment books"
authors = ["BigPigMoon matvey.novikov02@gmail.com"]
readme = "README.md"
packages = [{include = "book_worm"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.2"
python-dotenv = "^1.0.0"
peewee = "^3.16.2"
pylint = "^2.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
